Tests/LibWeb/Text/expected/css/animation-timing-function-per-keyframe.txt

linear at 0.25: 50px
linear at 0.5: 100px
ease at 0.25 is NOT linear: true
ease at 0.5: 160.484375px
per-keyframe linear at 0.25: 81.703125px
